convert fs-observable to typescript!
